pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "motadata-python-ai-sdk"
version = "0.1.0"
description = "Motadata Python AI SDK - Comprehensive SDK for building AI-powered applications"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Motadata", email = "info@motadata.com"}
]
keywords = ["ai", "llm", "rag", "agents", "sdk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "requests>=2.31.0",
    "redis>=5.0.0",
    "cachetools>=5.3.0",
    "aiofiles>=23.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "structlog>=23.0.0",
    "prometheus-client>=0.18.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "numpy>=1.24.0",
    "python-dateutil>=2.8.2",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/motadata/motadata-python-ai-sdk"
Documentation = "https://github.com/motadata/motadata-python-ai-sdk/blob/main/README.md"
Repository = "https://github.com/motadata/motadata-python-ai-sdk"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = false

# Ignore missing stubs for optional dependencies
[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "joblib.*",
    "mlflow.*",
    "PyPDF2.*",
    "speech_recognition.*",
    "pydub.*",
    "PIL.*",
    "pytesseract.*",
    "cv2.*",
    "bs4.*",
    "sentence_transformers.*",
    "src.faas.core.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

